teachme-rest-api/requirements.txt
alembic==0.9.7
certifi==2018.1.18
chardet==3.0.4
click==6.7
consulate==0.6.0
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Bcrypt==0.7.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-editor==1.0.3
requests==2.18.4
slerp-py==0.0.18
Pillow==5.0.0
PyPDF2==1.26.0
opencv-python==3.4.1.15